refactor(processing): split symbolic.rs into per-concern modules
symbolic.rs (1635 LOC) carried the 2D symbolic-annotation extractor: a
recursive item dispatcher plus per-primitive concerns. Convert it into a
symbolic/ directory:

  mod.rs        extract_symbolic_data orchestrator + `pub use` re-exports
  primitives.rs serde DTOs (SymbolicPolyline/Circle/Text/FillArea/GridAxis/Data)
  transform.rs  Transform2D core + placement parsers + circle_center
  items.rs      extract_symbolic_item dispatcher + trimmed-curve
  text.rs       extract_text_literal
  fill.rs       annotation fill-area + curve-ring sampling
  grid.rs       grid extraction + bubble emission
  color.rs      styled-item index + colour resolution

Pure code motion. Cross-module shared items (Transform2D + fields, the ~6
dispatcher/builder entry fns) gain pub(super); the rest stay private. The
six public DTOs + extract_symbolic_data keep their exact crate::symbolic::*
paths via `pub use`; lib.rs and types/response.rs resolve unchanged.
Submodules named primitives/color to avoid clashing with the crate-level
types/ and style/ dirs. cargo test -p ifc-lite-processing green (21 binaries);
clippy -D warnings clean; wasm32 build clean.
